Menu: cull non-visible items when `contentItem` inherits Item
For a custom menu control, when the contentItem does not inherit from
ItemView, the text overlaps. A related bug is QTBUG-53262, but this
problem is only fixed for the ItemView control.

Since the contentItem control can accept the Item control, I think the
repair only for ItemView has limitations. Therefore, this submission
cancels the restriction on the type of contentitem.

Pick-to: 6.3 6.4
Fixes: QTBUG-104470
Change-Id: I7e697e3cb993475d157a92f37d9512e4a9872387
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>