Fix fractional scaling of text in Qt Quick
As opposed to the raster engine, in Qt Quick we are using
unscaled positions for the glyphs and using the vertex shader
to scale these after the fact. However, when picking the
correct subpixel rendering for each glyph, we would use the
unscaled position's fractional part, meaning that we
essentially rendered the glyphs at the wrong subpixel position.
This was especially visible when doing fractional scaling, e.g.
125%.

Instead, we need to ensure that we pick the rendering at the
on-screen subpixel position.

[ChangeLog][QtQuick][Text] Fixed a kerning issue with
native-rendered text when there is a fractional system scale
factor set.

Fixes: QTBUG-101008
Change-Id: Ic0f94d8b4ca5998dca638bdf7e2a16306d92a926
Reviewed-by: Tor Arne Vestbø <tor.arne.vestbo@qt.io>
(cherry picked from commit 97cc7023ce8b8d903351a2e13f515592854c56fc)