Delete transition when it's canceled
If next transition occurs before previous transition ends, the
animation of the previous transition is canceled. The problem is
that it is not removed from A while it is being canceled.
At this time, if a new job comes in translator->runningJobs,'busy'
state cannot return back to its original state and mouseEvent/
touchEvent are broken.
So, in case of transition canceled, we should remove the transition.

Pick-to: 6.2 6.3 6.4
Fixes: QTBUG-104491
Change-Id: I33083822763bdc006eb35262b9c216f27989139a
Reviewed-by: Bumjoon Park <bumjoon.park@qt.io>
Reviewed-by: Richard Moe Gustavsen <richard.gustavsen@qt.io>