Repeater: uncull items that come from ObjectModel
After 6173fc0e4be93604286b9a424f1aa3ff5896082a, Menu will
unconditionally cull items as they are inserted. This behavior relies
on the fact that ListView will uncull the items as necessary. However,
when implementing Menu with a Repeater as the contentItem (which is
unusual but not unreasonable), the items will never be unculled and
hence will be invisible.

We can't fix this in Menu by checking its contentItem, because when
items are inserted, their parent item is the root contentItem, not the
Repeater, and there's no nice, efficient way to check if the item
belongs to a Repeater.

This patch fixes the issue in Repeater by unculling the item if it
came from an ObjectModel.

Done-with: Mitch Curtis <mitch.curtis@qt.io>
Fixes: QTBUG-104544
Change-Id: Ib37d2d49d3608156a3f1e3652b63ccdd5dfd0f41
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>
(cherry picked from commit 19637c2241724cac55315afbb70b5f4333a2065b)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>