QQuickMessageDialogImpl: make 'text' and 'informativeText' wrap
This is a behavior change.

Previously, the text or informativeText propertes were set to a very
large string without any newline characters, the dialog would increase
its implicit width in order to contain the whole line of text inside of
it.

This patch changes the behavior so that the implicit width of the
dialog will no longer care about the text and informativeText
properties, but rather wrap the text if it can't fit inside the dialog
with its natural text width.

This patch might also change the visuals of the dialog slightly.
Screenshots will be posted in the reported bug on JIRA.

I had some issues getting text wrapping to work while the Labels were
inside a ColumnLayout. Because of this, I decided to simply use Column
instead. This use case doesn't need the additional feature that a layout
would offer anyways.

Fixes: QTBUG-104269
Change-Id: I2efa7a697dfe5bdbb5d25ef2482a0581464013dc
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>
(cherry picked from commit 589d83bc9bc1da32a4f26934c54075ddb78b47b9)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>