Flickable: don't grab on press if already moving
9f9ea6e1837620a0ff4b98e262cba2df44215967 was too drastic: we cannot
simply avoid calling filterPointerEvent() on press, because it calls
handlePressEvent() which calls maybeBeginDrag() which sets pressPos,
among other things. Failing to set pressPos caused a regression:
flicking a second time while the flickable is already moving caused it
to "start over" as if it was flicking from the top again. So now we
revert that fix, and instead avoid setting stealMouse to true in
handlePressEvent(). This makes the second flick more like the first:
it waits for move events before taking the exclusive grab; so it's ok
for a delegate (such as a Button or MouseArea) to detect the press and
take a grab in the meantime. Flickable does not need to grab on press
during filtering: it's not a lost opportunity, because it keeps
filtering its children's events later on anyway.

The flickDuringFlicking() test is added for this scenario. For the same
reason as tapDelegateDuringFlicking(), this is also likely to fail on
Android for now.

This could have some impact on d7b5a485583004ad6a1374a50b7c3f6cab00aca3
but it seems the test that was added there still passes.

Fixes: QTBUG-103832
Task-number: QTBUG-38765
Task-number: QTBUG-74842
Task-number: QTBUG-104471
Pick-to: 6.2 6.3 6.4
Change-Id: Icb45fcd94847051121ee78a970fbd5f5dc8a42a4
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Antti Määttä <antti.maatta@qt.io>
Reviewed-by: Richard Moe Gustavsen <richard.gustavsen@qt.io>
Reviewed-by: Oliver Eftevaag <oliver.eftevaag@qt.io>