QmlCompiler: Clean up state merging code
Generally, we need to merge all registers when merging different
branches of execution in the type propagator, even those that were
written by the same instruction on all branches. Doing this right
relieves us from the need to special case the arguments (which most of
the time are only "written" once) all over the place.

Pick-to: 6.4
Task-number: QTBUG-104462
Change-Id: If30f7382c43d1e04ad91c88cfbc217e070de518a
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>