qmltyperegistrar: Adjust to QMetaTypeInterface no longer being mutable
We must not create a const QMetaTypeInterface variable, as we are going
to modify it shortly after by calling id on it.
On the other hand we can make it Q_CONSTINIT, as metaTypeForNamespace is
constexpr.

Adapts to c474825945a418ae9d8494c80b6803b6279e0cf5 in qtbase. The change
in qtbase keeps being BC: Generated code (by qmltyperegistrar) will
either use the new code where the interface is no longer const, or it
has been compiled against a qtbase where the member was still mutable.
We do not support compiling the generated code against a different Qt
version.

Change-Id: I3ba285ff17b4b3fd51ce1db4a307ff78289cae45
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>