QQuickWindow: Implement QWindowPrivate::clearFocusObject()
Change 68c103225f4e8bd removed the overridden function
QQuickWindowPrivate::clearFocusObject(). This is quite
unfortunate, since it's used by e.g QPA/iOS to clear
the focus object when the input panel is closed.

The result is that if you hide the input panel by
swiping it down, a TextField with focus will not loose
focus anymore, but continue to show a blinking cursor.
Moreover, you cannot bring the input panel up again by
clicking on the TextField, since it's focused already, so
no change will be reported back to the platform IM.

Pick-to: 6.4 6.3 6.2
Fixes: QTBUG-104642
Change-Id: Ic27aaa2cc8d11cbf64ba8d3797ab4008bc2eb3a6
Reviewed-by: Tor Arne Vestbø <tor.arne.vestbo@qt.io>