macOS: ensure FocusFrame stays on top of the control, also when using layouts
FocusFrame is parented to have the same parent as targetItem.
This is done so that the drawn focus frame stays on top of the
target item's siblings, to avoid "clipping" glitches.

But when the parent is a layout, the layout will actually lay
out the FocusFrame as well, and position it somewhere else than
the expected 0,0. This, again, will affect the relative position
of the inner Rectangle (that actually draws the frame), so that
it ends up at a wrong position.

To ensure that the FocusFrame stays on top of the target item, also
when it's position is different than 0,0, we need to subtract root's
position from the position of the Rectangle.

Fixes: QTBUG-102559
Change-Id: If266cc0cf6cea50834fd4db5e2a06a0d31394070
Reviewed-by: Doris Verria <doris.verria@qt.io>
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>
(cherry picked from commit 3ed72de5e211b2b97561ffc6ac69320404a0cb2f)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>