Purge QQmlBindingBinding
This patch removes QQmlBindingBinding, as well as QQmlNonBindingBinding
as a standalone class. The functionality of QQmlNonBindingBinding is
merged into QQmlBinding.

QQmlBindingBinding was used to store a QML binding in a property;
the only user of it was Binding's when property, and that was removed in
Qt 5.14 already.
The big disadvantage of its approach was that it has zero type safety,
and interacts in weird ways with bindings created via Qt.binding; it
also does not integrate well with QProperty/C++ based bindings.
Lastly, it is completely untested.

As QQmlBinding is a private class, we can safely remove the whole
functionality, as there is no Qt internal user of it anymore.

Use cases which require the manipulation of the binding should instead
use a well-typed property, and then access the binding via the
QQmlAnyBinding or QQmlAbstractBinding API (compare Binding's when
property in Qt 6.2+).

Change-Id: I6de0c9987d371a04c7f78047bb34bf684b6a69f5
Reviewed-by: Andrei Golubev <andrei.golubev@qt.io>