Ensure that when an item resizes, its layer resizes immediately
geometryChange() notifies QQuickItemLayer::updateGeometry() to update
its internal bounds; at that moment, QQuickImagePrivate::paintedHeight
has not yet been updated, so the QQuickImage::boundingRect() override
returns the old size. It gets updated in updatePaintedGeometry(); but
until then, it's better to use QQuickItem::boundingRect() without
polymorphism, because that simply returns the item width and height.
An alternative would be to use clipRect(), as it was before
9b62f4c27ac3fb3dc563c7f4657094c14d752bac; but that is more complex now:
in case ItemObservesViewport is set, we would need to worry about
whether it's more appropriate to have the item's layer the same size
as the viewport (and update it more often during scrolling), or the
same size as the item, as it was before.

Amends 9b62f4c27ac3fb3dc563c7f4657094c14d752bac

Fixes: QTBUG-104442
Fixes: QTBUG-104536
Pick-to: 6.4 6.3
Change-Id: I51e25402d358a57cea9fd718ee8fe759b572b1a2
Reviewed-by: Igor Bugaev <freedbrt@gmail.com>
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>