QQuickLabsPlatformFolderDialog: Modify the logic of setting the folder
In QQuickPlatformFolderDialog, when triggering acceptance, the folder is
set by the value of currentFolder, and the value is set back to
currentFolder inside the setFolder function. This action causes
currentFolder to be abnormal after the folder is modified.

Fixes: QTBUG-104629
Change-Id: I99978de5bbd019024c499c50f61d6e7b6e7dc140
Reviewed-by: Oliver Eftevaag <oliver.eftevaag@qt.io>
(cherry picked from commit 5e8b7627565d408ed6dbd9f3a192d0bbeff6720e)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>