Native style: ensure implicitSize is not less than minimumSize
After e0990f8b406, we now send a content rect that is smaller than
the minimum size to QQuickMacStyle::sizeFromContents(CT_LineEdit).
Because the affected code in sizeFromContents doesn't expect that,
it returns a size that is too small to fit the background.

Since a QQuickStyleItemTextField will respect the minimum size, and
therefore end up drawing a background that is bigger than the
control, the focus rect (that respects the size of the control) will
appear too big.

This patch will ensure that sizeFromContents never returns a size
that is smaller than the minimum size, even if the contents are
smaller. A small check in the code for this is also added, so that
we can avoid similar cases early on later.

Fixes: QTBUG-104705
Pick-to: 6.4 6.3
Change-Id: Ifb763dcce5cef43e6aec9a7da70d73bf8f1255ff
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>