macOS, SpinBox: Let SpinBox use TextField internally
By letting a SpinBox on macOS use a TextField as contentItem (as
opposed to TextInput), we get all the styling and sizing for
free. This will also fix a minor issue where the focus frame
didn't align perfectly to the input.

In the same go, remove all the extra code in SpinBox.qml that
handles the case where the application customizes it - it will
never look correct anyway, and just complicates the code.

Change-Id: Ie4a675a85c6b9535a69b28d65422abae60d3f26d
Reviewed-by: Doris Verria <doris.verria@qt.io>