macOS, ScrollBar: implement an opacity of 0.5
The current look of the ScrollBar is far to dark on macOS.
From comparing with Finder, an opacity of 0.5 looks to give
the exact same shade of gray. The scrollbar on macOS should
also be semi-transparent, to let the bg color of the window
shine through.

Change-Id: Ic6df18fcbb6a1c6c5e9a31d40188f320a965a6d2
Reviewed-by: Doris Verria <doris.verria@qt.io>
(cherry picked from commit 002f24c06ffddaf7fff3655d090d343692a412e5)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>