Native style, StyleItemScrollbar: don't draw the scrollbar as sunken when using an override state
When we specify an override state from QML, like NeverHovered, we
also don't want the scrollbar to be drawn as pressed (sunken).
The point with the override state is to lock the image to a
certain state that doesn't change.

This will fix a drawing issue on macOS, where the scrollbar
becomes far too dark when you drag on it.

Change-Id: I7f62e7ebd6d1ebbe808e035f63485ed292661bc1
Reviewed-by: Jan Arve Sæther <jan-arve.saether@qt.io>
(cherry picked from commit 5a8ac8f3d57c5381afb6a4d79990b0d7a5298936)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>