tst_qqmllanguage: Adjust to QMetaTypeInterface no longer being mutable
We must not create a const QMetaTypeInterface variable, as we are going
to modify it shortly after by calling id on it.
On the other hand we can make it Q_CONSTINIT, as metaTypeForNamespace is
constexpr.

Adapts to c474825945a418ae9d8494c80b6803b6279e0cf5 in qtbase.
Amends commit 5db7944de8440087d31b76e6fb76c463e7c00490.

Change-Id: Idbf044b676b6f3859198aeefcb0322e9e9461e5c
Reviewed-by: Sami Shalayel <sami.shalayel@qt.io>
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>