QQuickFontDialog: make selectedFont writable before open() is called
This was an oversight, that I had forgotten to fix earlier.
The dialogs in Qt Quick Dialogs are created lazily, upon the first call
to open().
If the selectedFont property was written to before the dialog had
been created from the first call to open(), then the setter would
simply return early, and not do anything.

This change guarantees that the selectedFont will be set, regardless of
whether it's been opened or not.

This change has already been done for the FileDialog and the ColorDialog.

Change-Id: Ic8906a197931e64bb7a3d21aa9f9ca36ba5fd840
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>
(cherry picked from commit 843b153e6e78095a0bea839f7cbb4e09c47ea3b2)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>