Rename Test module from tst_qquicklistview2
The CI mismatches the module Test defined in
tests/auto/quick/qquicklistview2/CMakeLists.txt
with the another one with the same name defined in
tests/auto/qml/qqmllanguage/CMakeLists.txt

Renamed Test to TestQQuickListView2 as a workaround. Renaming the module
from qqmllanguage would be bothersome as it is used in a lot of files.

Fixes: QTBUG-104698
Pick-to: 6.2
Change-Id: Ia0107e8c3ce4e4d386ecd5bbae56953b3a9abc8b
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>