QtQuickDialogs: Document the existence of the DontUseNativeDialog option
There are 2 different methods that can be used to force the dialogs in
the QtQuickDialogs module to use the non-native implementation.

The first method is to set the AA_DontUseNativeDialogs application
attribute, which forced all dialogs to be non-native globally.

The second option is to change the value of the options property for the
specific dialog that one wish to be non-native, to a value where the
DontUseNativeDialog flag is set.

This is an enum value that should be mentioned in the documentation.

Fixes: QTBUG-99349
Pick-to: 6.4
Change-Id: Icc3aca7cb3dbfeb6c01a2e9d382a58cfcc2dc84a
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>