sg: Prevent not supported format warnings when wrapping external textures
Amends 1df2cf6bad7207f16ddca17344cc1324e50f287e

Clearly the intention is to continue using RGBA8 as the default when
nothing is specified, but then the unknown/invalid value is not handled
correctly in the conversion functions.

As a result, when running f.ex. tst_qquickrendercontrol one gets
warnings such as "VkFormat 0 is not supported". This is now avoided.

Pick-to: 6.4
Change-Id: I0dd2cda3fefe24a1bdac075892df5a6eb8796181
Reviewed-by: Andy Nichols <andy.nichols@qt.io>