Introduce QQml::makeRefPointer helper function
This creates a QQmlRefPointer which owns a newly allocated T. This
ensures that we use the correct mode when creating the QQmlRefPointer.
It's usage is similar to make_shared and make_unique, but it requires T
to publically derive from QQmlRefCount (so that we can be sure that
Adopt is the correct usage mode).

Change-Id: I66f1c918a3249d2e28e1ac9e3d84baa839a4c253
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>
Reviewed-by: Sami Shalayel <sami.shalayel@qt.io>