QQDAPrivate::deliverUpdatedPoints: skip HoverHandlers if mouse event
Since bbcc2657fa0dbf715e6db7d675662e4be94a1e04,
QQuickDeliveryAgentPrivate::deliverHoverEventToItem() delivers synthetic
QMouseEvents to all HoverHandlers that need it; and when it encounters
any that has its blocking property set, it skips the rest. Then
QQDeliveryAgentPrivate::deliverUpdatedPoints() visits any remaining
handlers that did not yet get a chance to see the event; that should
skip HoverHandlers, because deliverHoverEventToItem() already visited
those that should be visited, and the remaining ones should not be.
But this mechanism is not quite the same with QTabletEvents:
QQuickDeliveryAgentPrivate::deliverHoverEvent() is not suitable,
because we want to send the QTabletEvent rather than a synth-mouse;
so if the event is not a mouse event, it's ok for deliverUpdatedPoints()
to deliver it to a HoverHandler.

Pick-to: 6.3 6.4
Task-number: QTBUG-101932
Change-Id: Ifb61d3e0b097f19a14595726afcf1f2e4be84f3f
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>