Add explicit dependency on 'Qml_sync_headers' for QML modules
Add dependency on 'Qml_sync_headers' for all QML modules that are built
in the qtdeclarative build tree.

'moc' requires QtQml/qqmlextensionplugin.h to process IID in the
generated plugin source code. Since the '_autgen/timestamp' target
doesn't have the 'Qml_sync_headers' target in its dependency chain,
the header file is not created by syncqt before the 'moc' run. We need
to make this explicitly, because CMake doesn't add the main target
dependencies to the dependency list of  the '_autgen/timestamp' target.

Change-Id: I1a199b4044e1dd9caddf7c0a9d75d59860893f60
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>