sg: Remove logic for creating oversized upload pools
When this was added in 2015, the default code path put vertex and
index data into the same buffer which also involved using the same
"pool" (not really a pool, just temporary backing storage) for
both. The guess is that the largest-vertex-buffer-size * 2 logic was
meant to accommodate for this, i.e. an attempt to reduce the chance of
further resize() calls later on by setting a hopefully-large-enough
size up front, while doubling the size in an attempt to end up with a
QDataBuffer large enough for largest vertex+index combination.

In Qt 6 (or in Qt 5 in some cases, such as with the WebGL platform
plugin), vertex and index buffers are always separate. Making both of
these "pools" to have a size of largest * 2 upfront is wasteful, they
will never need to store more data than the largest buffer.

Change-Id: Iaad9da81e6692fad2458a1f8414787239821ef59
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>
Reviewed-by: Andy Nichols <andy.nichols@qt.io>