Avoid use-after-free in QQuickPathView
When releasing an item we always need to unregister the change listener,
even if the model is already gone. The model is a public property and a
QPointer. Anything can happen to it behind our back.

Furthermore, we apparently don't own items and itemCache. Therefore, we
have to listen to their "destroy" events and remove them from the list
when those are triggered. Otherwise we operate on dangling pointers
when we later releaseItem() from the dtor.

Pick-to: 6.2 6.3 6.4
Change-Id: I662b0d70b8b06d15b504c6fe1fc33384b1ae12d9
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>