Android: Fix crash on tap handler with a S-Pen
This patch fixes a crash using a S-Pen with QML TapHandler on Qt 5.15.
The reason it was happening was because the item->window() was nullptr
when it tried to run the mapToGlobal function. The item->windows becomes
nullptr on handlePointerEvent function, so switching both lines is
enough to fix this crash.

Fixes: QTBUG-86805
Change-Id: I25f20650e97b5ad094a99c810379fb412fdb4222
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>