tst_qquickmenu: Fix click position and wait for menu opened() signal
- When clicking on the menu, take into account the menu's coordinates
as this may not always be (0,0) depending on the style-specific margins.

- Wait for the menu enter transitions to finish before trying to
interact with it, otherwise it leads to flakiness.

- Explicitly set the overlap property to 0 when testing the submenus
as the test relies on the menuItem to be clickable and not hidden by
an overlapping subMenu item.

Change-Id: I561f47e69b2c9ee12c46f6accacd06e22b448df5
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>
(cherry picked from commit 76d992431e113c3df9b6f879e2889973d5093a44)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>