macOS: Add Dialog and DialogButtonBox
This fixes the issue where Fusion style Buttons were displayed
by DialogButtonBox for its standardButtons.

We need to provide our own implementations of Dialog and DialogButtonBox
to ensure that the correct files get picked up.

The implementations themselves are almost identical copies of the Fusion
style's, with some imports removed.

The Fusion style is the macOS style's fallback.

Fixes: QTBUG-104658
Pick-to: 6.2 6.3 6.4
Change-Id: Ib7958565ee62f6f4286f87298ab98209e916b6b0
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>