QQuickItem: make map(From|To)(Item|Global) type-safe in QML
Instead of exposing a QQmlV4Function* overload, we can expose an
overload set to the engine. We do not gain anything from using
QQmlV4Function*, as we otherwise would need to manually parse the
arguments we got and dispatch to the correct internal function.

Unfortunately, we can only deprecate the QQmlV4Function* overload, not
outright remove it: While no public API allows creating a
QQmlV4Function* object, and passing a null pointer would lead to an
immediate crash, it is possible to write a Q_INVOKBALE that forwards its
argument to one of the existing functions. That could then be used in
QML. Thus, we can only deprecate the function in Qt 6.

We make the new overloads public to ensure that the methods get found by
the engine even after the engine ignores private invkoables (compare
QTBUG-105136).

Task-number: QTBUG-105139
Change-Id: I6af2c69f4bd2b70ae45936138808313421085b01
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>