Flipable: switch sides at 90 degrees regardless of camera distance
QQuickRotation::applyTo(QMatrix4x4 *matrix) uses
QMatrix4x4::projectedRotate() to apply rotation in pseudo-3D.
QQuickFlipablePrivate::updateSide() was previously applying
QQuickRotation to QTransform to decide which side should be showing;
but when using QTransform::map(QPointF) to get the position of the
transformed point, it is affected by the distance-to-plane value.
QMatrix4x4::projectedRotate() uses 1024 as distance from "camera" to
plane; so if the items on the plane are too large, some coordinates on
the plane may be behind the camera during rotation, which could cause
QQuickFlipable to mistakenly think that the current side has become
QQuickFlipable::Back.

Now we directly calculate the side of QQuickFlipable in 3D space rather
than performing projection, to avoid being affected by the camera
distance value. We use QMatrix4x4 instead of QTransform.

Fixes: QTBUG-75954
Pick-to: 6.5 6.4 6.2
Change-Id: Ibb08f6f6de690f5b56c3130a1f7b09beb250c807
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
Reviewed-by: Mikołaj Boc <Mikolaj.Boc@qt.io>
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>