ScrollView: ignore filtered wheel events when wheelEnabled is false
The event filter was returning true when wheelEnabled was false,
but wasn't ignoring the event, meaning that propagation of the wheel
event stopped there, and items behind the ScrollView didn't get the
event even if ScrollView wasn't using it.

I'm not sure why it was done this way - the behavior isn't documented
and no auto tests fail when changing it.

Fixes: QTBUG-105164
Change-Id: Ie826ccfc406b77606ef0c6f043dc48f42a18cdc7
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>
(cherry picked from commit 7ef12715baf48912f713e6cb3ee2a784caaadaae)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>