Use the PlatformServiceColorPicker for the ColorDialog's eye dropper
On wayland, it's not possible to use QScreen::grabWindow(), in order to
to pick a color from the screen.
However, it is possible to use the XDG desktop portal, to pick a color
for us.

In qtbase the commit 00a8957cef8f9406346d299da03f595427078f43 added a
new virtual function pickColor() to QPlatformService.

This patch takes advantage of this new API, to prioritize using this new
platform service if available, and fallback to the old behavior in case
it's not supported for the current platform.

Change-Id: Ia155531ff4e6ee1446f122dcb44d90753c11846b
Reviewed-by: Tor Arne Vestbø <tor.arne.vestbo@qt.io>