QQuickLoader: Remove QQmlV4Function usage
Just add a normal overload set. Unfortunately, one overload would
conflict with the plain setter of the source property; but we need a
difference in behavior: The plain setter should not resolve the URL,
whereas the invokable should.
Solve this by renaming the setter (can be done, as the type is private).

Task-number: QTBUG-105139
Change-Id: I27db968935f6eeddd415ccc6cb8c06ad6f0cf896
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>